mcp-server/requirements.txt
mcp[cli]
psycopg2-binary
pgvector
uvicorn
sentence-transformers
numpy
